Crate clippy_tracing

A tool to add, remove and check for tracing::instrument in large projects where it is infeasible to manually add it to thousands of functions.

Macros

create_check_visitor_function ğŸ”’

create_fix_visitor_function ğŸ”’

create_strip_visitor_function ğŸ”’

Structs

CheckVisitor ğŸ”’

Visitor for the check action.

CommandLineArgs ğŸ”’

The command line arguments for the application.

Desc ğŸ”’

The description of attributes on a function signature we care about.

FixVisitor ğŸ”’

Visitor for the fix action.

SegmentedList ğŸ”’

A list of text lines split so that newlines can be efficiently inserted between them.

StripVisitor ğŸ”’

Visitor for the strip action.

Enums

Action ğŸ”’

The action to take.

ApplyError ğŸ”’

Error for apply.

ExecError ğŸ”’

Error for exec.

Exit ğŸ”’

Type to return from main to support returning an error then handling it.

Functions

apply ğŸ”’

Apply the given action to the given source and outputs the result to the target produced by the given closure.

check_attributes ğŸ”’

Returns a tuple where the 1st element is whether tracing::instrument is found in the list of attributes and the 2nd is whether clippy_tracing_attributes::skip is found in the list of attributes.

exec ğŸ”’

Wraps functionality from main to support returning an error then handling it.

find_instrumented ğŸ”’

Finds the #[instrument] attribute on a function.

inner_instrument ğŸ”’

Returns the instrument macro for a given function signature.

instrument ğŸ”’

is_instrumented ğŸ”’

Checks if a syn::Attribute is #[instrument].

main ğŸ”’